OLX Car Selling Data Analysis and Visualization
Hi All,
I have worked on a CSV that is related to the OLX Car Selling website database, the data is from Year 1995 to Year 2020.
I used the Python as Programming Language with different popular Libraries i.e. Pandas, Matplotlib, Seaborn.
The Data cleaning is performed through Pandas and Visualization are performed with the help of Matplotlib and Seabron.
I have extracted the different patterns and information from the data and tried to visualize these by using different types of plot and charts.
Hope it will helpful for you.